Switch stage's type to int from char
Char was just giving trouble during type conversion from string during
parsing the input file as well as have had to cast to int for screen
output.

+++ src/Control/Keywords.h
@@ -1487,7 +1487,7 @@ struct nstage_info {
     Issue 3, 2000, Pages 249-275.)";
   }
   struct expect {
-    using type = uint8_t;
+    using type = unsigned int;
     static constexpr type lower = 2;
     static std::string description() { return "uint"; }
   };
